Move FrameRateVotingManager to separate file
